Corrección box_loss

+++ CustomFaster_MP_TPU/rpn.py
@@ -347,7 +347,7 @@ class RegionProposalNetwork(torch.nn.Module):
         class_weights = torch.tensor([0.5, 2]).to(device)
 
         
-        sampled_classes = labels[sampled_pos_inds].to(device)  
+        sampled_classes = labels[sampled_pos_inds].long().to(device)  
 
         # Obtener los pesos correspondientes
         sampled_weights = class_weights[sampled_classes].to(device)
